Scripting of key imports on PDP and changing SQL
Added a function to the keys.sh to support import PEP key files
onto to PDP. This was tested on OM2 with a breakout script and
its changes merged in.

Updated pep and pdp scripts to support this new function.

Added functions to keys.sh to support PDP generation and
importing (locally) of MySQL keys. Also commented the
MySQL key import blocks as they were confusing even me.
The comments are pointing to  support-docs/mysql-keypairs.txt
section headers.

Finally there is a new script that should ONLY be run in
this directory structure to prepare the users.sql file from the
PDP for importing. It changes all the IP address, netmasks and
passwords of the file - which used to have to be done manually.